feat(ec2): add support for NAT instances
Add support for NAT instances (as opposed to NAT gateways)
on VPCs. Introduced the concept of a 'NAT provider', and provide
two implementations: one for gateways, one for instances.

To make it easier to pick the right AMI for the NAT instance,
add an AMI context provider, which will look up AMIs available to
the user.

Fixes #4876.